pyproject.toml
[project]
name = "get-opinions"
version = "1.0.0"
description = "OPINIONS — RAG-based opinion retrieval from indexed personas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cli2>=5.2",
    "exa-py>=2.4.0",
    "firecrawl-py>=4.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "symbolica-agentica>=0.4.1",
]

[project.scripts]
opinions = "main:main"

[tool.setuptools]
py-modules = ["main", "pipeline", "tools", "personas", "agents"]
